refactor based on abn chagnes
