+++ abn/watcher/informer_test.go
@@ -3,14 +3,14 @@ package watcher
 import (
 	"testing"
 
-	"github.com/iter8-tools/iter8/driver"
+	"github.com/iter8-tools/iter8/base"
 	"github.com/stretchr/testify/assert"
 	"helm.sh/helm/v3/pkg/cli"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
 func TestNewInformer(t *testing.T) {
-	kd := driver.NewFakeKubeDriver(cli.New())
+	kd := base.NewFakeKubeDriver(cli.New())
 	w := NewIter8Watcher(
 		kd,
 		[]schema.GroupVersionResource{{
